feat: remove material-ui, componentize, loading / login
