+++ main.js
@@ -25,10 +25,7 @@ function createWindow () {
     width: 1024,
     height: 768,
     show: false,
-    frame: false,
-    webPreferences: {
-      webSecurity: false,
-    }
+    frame: false
   });
 
   // and load the index.html of the app.
