+++ src/assets/css/App.css
@@ -1,3 +1,7 @@
+* {
+  box-sizing: border-box
+}
+
 body, html {
   background: hsl(0, 0%, 10%);
   color: white;
@@ -9,6 +13,67 @@ body, html {
 
 #app {
   height: 100%;
+}
+
+/* Typography */
+html, body, input, button {
+  font-family: 'Roboto';
+  line-height: 1.6;
+}
+
+ul,
+ol {
+  margin-bottom: 1.53rem;
+}
+
+h1 {
+  font-size: 2.4414rem;
+  line-height: 2.7rem;
+  margin-bottom: 1.53rem;
+}
+
+h2 {
+  font-size: 1.9531rem;
+  line-height: 2.7rem;
+  margin-bottom: 1.53rem;
+}
+
+h3 {
+  font-size: 1.5625rem;
+  line-height: 1.8rem;
+  margin-bottom: 1.53rem;
+}
+
+h4 {
+  font-size: 1.25rem;
+  line-height: 1.8rem;
+  margin-bottom: 1.53rem;
+}
+
+h5 {
+  font-size: 1rem;
+  line-height: 1.8rem;
+  margin-bottom: 1.53rem;
+}
+
+h6 {
+  font-size: 1rem;
+  line-height: 1.8rem;
+  margin-bottom: 1.53rem;
+}
+
+body {
+  font-size: 1rem;
+  line-height: 1.8;
+}
+
+p,
+ol,
+ul,
+pre {
+  font-size: 1rem;
+  line-height: 1.8;
+  margin-bottom: 1.53rem;
 }
 
 h1 {
@@ -23,6 +88,19 @@ a {
 
 a:hover {
   color: #FFF
+}
+
+/* Helpers */
+.text-center {
+  text-align: center;
+}
+
+.text-right {
+  text-align: right;
+}
+
+.text-left {
+  text-align: left;
 }
 
 /* Paper */
@@ -48,12 +126,13 @@ a:hover {
   -webkit-app-region: drag
 }
 
+a,
 button,
 .tab-selector,
-.app-control,
-.app-control div,
-.account-actions,
-.not-draggable {
+.app-control *,
+.account-actions *,
+.not-draggable,
+.not-draggable * {
   -webkit-app-region: no-drag
 }
 
@@ -107,6 +186,25 @@ button,
 
 .app-control .fullscreen-control:hover {
   background: #81c784
+}
+
+/* Main Viewport */
+.viewport {
+  height: 100vh
+}
+
+.main-viewport {
+  position: fixed;
+  padding: 15px;
+  top: 94px;
+  left: 0;
+  right: 0;
+  height: calc(100vh - 94px);
+}
+
+/* App Viewport */
+.app-viewport {
+  margin: 1px;
 }
 
 /* Tabs */
