+++ src/assets/css/login.css
@@ -1,0 +1,61 @@
+.login-viewport {
+  background-size: cover;
+  background-image: url('../background.png')
+}
+
+.login-viewport .viewport {
+  background: rgba(0, 0, 0, 0.8);
+}
+
+.login-container {
+  width: 300px;
+  /* background: #222; */
+  border-radius: 3px;
+  padding: 20px;
+  /* box-shadow: 0 10px 10px rgba(0,0,0,0.5); */
+  /* color: #222; */
+}
+
+.login-container .error {
+  font-size: 12px;
+  padding: 8px 12px;
+  background: hsla(0, 100%, 24%, 0.5);
+  border-radius: 8px;
+  border: 1px solid hsla(0, 80%, 9%, 0.8);
+  box-shadow: inset 0 0 1px hsla(0, 100%, 100%, 0.6), 0 0 10px rgba(0,0,0,0.3);
+}
+
+.login-container label {
+  font-size: 14px;
+  font-weight: 500;
+}
+
+.login-container .form-input {
+  position: relative;
+  border: none;
+  color: #222;
+  background: white;
+  border-radius: 4px;
+  padding: 8px;
+  cursor: pointer;
+  transition: all .3s;
+  outline: none;
+  margin-bottom: 8px;
+  width: calc(100%);
+  border: 1px solid rgba(255,255,255,0.1);
+}
+
+.login-container .form-input + p {
+  font-size: .75rem;
+  margin-top: 0;
+  color: #888;
+  text-align: center;
+  /* margin-bottom: 0; */
+}
+
+.login-container .btn-primary {
+  width: 80%;
+  margin: 7px auto 32px;
+  display: block;
+  /* padding: 12px 8px; */
+}
